+++ tests/examples/amend_delete_both/main.sh
@@ -4,16 +4,19 @@ set -e
 trap 'kill $(pgrep -g $$ | grep -v $$) > /dev/null 2> /dev/null || :' EXIT
 rm -rvf $(cat .gitignore)
 
-# Run the example
-echo "created elsewhere" > test1.txt
-stepup -w -n 1 plan.py & # > current_stdout1.txt &
+# Create input files
+echo "This is the first half and" > asource1.txt
+echo "this is the second half." > asource2.txt
+
+# Run the plan.
+stepup -w -e -n 1 plan.py & # > current_stdout1.txt &
 
 # Wait for the director and get its socket.
 export STEPUP_DIRECTOR_SOCKET=$(
   python -c "import stepup.core.director; print(stepup.core.director.get_socket())"
 )
 
-# Get the graph after completion of the pending steps.
+# Initial graph
 python3 - << EOD
 from stepup.core.interact import *
 wait()
@@ -26,26 +29,23 @@ wait
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1
-[[ -f test1.txt ]] || exit 1
-[[ -f test2.txt ]] || exit 1
+[[ -f asource1.txt ]] || exit 1
+[[ -f asource2.txt ]] || exit 1
+[[ -f data1.txt ]] || exit 1
+[[ -f data2.txt ]] || exit 1
+[[ -f log.txt ]] || exit 1
 
-# Call cleanup to remove all outputs
-cleanup -vv . > current_cleanup.txt
-
-# Check files that are expected to be present and/or missing.
-[[ -f plan.py ]] || exit 1
-[[ -f test1.txt ]] || exit 1
-[[ ! -f test2.txt ]] || exit 1
-
-# Restart without changes
-stepup -w -n 1 plan.py > current_stdout2.txt &
+# Remove both intermediates and restart
+rm data1.txt data2.txt log.txt asource1.txt
+stepup -w -e -n 1 --log-level INFO plan.py & # > current_stdout2.txt &
+PID=$!
 
 # Wait for the director and get its socket.
 export STEPUP_DIRECTOR_SOCKET=$(
   python -c "import stepup.core.director; print(stepup.core.director.get_socket())"
 )
 
-# Get the graph after completion of the pending steps.
+# Initial graph
 python3 - << EOD
 from stepup.core.interact import *
 wait()
@@ -54,11 +54,13 @@ join()
 EOD
 
 # Wait for background processes, if any.
-wait
+set +e; wait -fn $PID; RETURNCODE=$?; set -e
+[[ "${RETURNCODE}" -eq 3 ]] || exit 1
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1
-[[ -f test1.txt ]] || exit 1
-[[ -f test2.txt ]] || exit 1
-# The file test1.txt was AWAITED, so changes to this file are not relevant for StepUp.
-[[ $(grep -c "UPDATED │ test1.txt" current_stdout2.txt ) -eq 0 ]] || exit 1
+[[ ! -f asource1.txt ]] || exit 1
+[[ -f asource2.txt ]] || exit 1
+[[ ! -f data1.txt ]] || exit 1
+[[ -f data2.txt ]] || exit 1
+[[ ! -f log.txt ]] || exit 1